Remove base-model
